refactor chat again
