--- crabml-llama2/src/chat/mod.rs
@@ -1,5 +1,0 @@
-mod chat;
-mod mark_matcher;
-
-pub use chat::Llama2Chat;
-pub use chat::Llama2ChatReplyIterator;
